+++ README.md
@@ -1,1 +1,16 @@
+Structure of the project
 
+- Basic React App Layout
+-
+- Header
+- - Logo
+- - Navigation
+- BODY
+- - Search Bar
+- - Restaurant Contianer
+- - Restaurant Cards
+- FOOTER
+- - Contact
+- - Copy Right
+- - Social Media
+- - Links
