+++ src/components/BodyComponent.js
@@ -1,0 +1,2174 @@
+import RestaurantCardComponent from "./RestaurantCardComponent";
+
+const resObjArray = [
+  {
+    "info": {
+      "id": "336917",
+      "name": "Pizza Hut",
+      "cloudinaryImageId": "2b4f62d606d1b2bfba9ba9e5386fabb7",
+      "locality": "Haibatpur Road",
+      "areaName": "Crossing Republic",
+      "costForTwo": "₹350 for two",
+      "cuisines": [
+        "Fast Food",
+        "Pizzas"
+      ],
+      "avgRating": 3.8,
+      "feeDetails": {
+        "restaurantId": "336917",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3300
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3300
+      },
+      "parentId": "721",
+      "avgRatingString": "3.8",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 28,
+        "lastMileTravel": 2.4,
+        "serviceability": "SERVICEABLE",
+        "slaString": "28 mins",
+        "lastMileTravelString": "2.4 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "30% OFF",
+        "subHeader": "UPTO ₹75"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/pizza-hut-haibatpur-road-crossing-republic-noida-336917",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "340381",
+      "name": "KFC",
+      "cloudinaryImageId": "56c9ab92bd79745fd152a30fa2525426",
+      "locality": "Sector 4",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹400 for two",
+      "cuisines": [
+        "Burgers",
+        "Biryani",
+        "American",
+        "Snacks",
+        "Fast Food"
+      ],
+      "avgRating": 4,
+      "feeDetails": {
+        "restaurantId": "340381",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3800
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3800
+      },
+      "parentId": "547",
+      "avgRatingString": "4.0",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 21,
+        "lastMileTravel": 2.4,
+        "serviceability": "SERVICEABLE",
+        "slaString": "21 mins",
+        "lastMileTravelString": "2.4 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "20% OFF",
+        "subHeader": "UPTO ₹50"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/kfc-sector-4-gaur-city-1-noida-340381",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "189174",
+      "name": "Burger King",
+      "cloudinaryImageId": "e33e1d3ba7d6b2bb0d45e1001b731fcf",
+      "locality": "Sector 4",
+      "areaName": "Sector 4",
+      "costForTwo": "₹350 for two",
+      "cuisines": [
+        "Burgers",
+        "American"
+      ],
+      "avgRating": 4,
+      "feeDetails": {
+        "restaurantId": "189174",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3900
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3900
+      },
+      "parentId": "166",
+      "avgRatingString": "4.0",
+      "totalRatingsString": "10K+",
+      "sla": {
+        "deliveryTime": 30,
+        "lastMileTravel": 3,
+        "serviceability": "SERVICEABLE",
+        "slaString": "30 mins",
+        "lastMileTravelString": "3.0 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 03:45:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "EVERY ITEM",
+        "subHeader": "@ ₹129"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/burger-king-sector-4-noida-189174",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "147910",
+      "name": "La Pino'z Pizza",
+      "cloudinaryImageId": "ux1ryy9yqhrqwznhl3ju",
+      "locality": "Mahagun Mart",
+      "areaName": "Gaur City 2",
+      "costForTwo": "₹400 for two",
+      "cuisines": [
+        "Pizzas",
+        "Pastas",
+        "Italian",
+        "Desserts",
+        "Beverages"
+      ],
+      "avgRating": 3.8,
+      "feeDetails": {
+        "restaurantId": "147910",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 4500
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 4500
+      },
+      "parentId": "4961",
+      "avgRatingString": "3.8",
+      "totalRatingsString": "5K+",
+      "sla": {
+        "deliveryTime": 27,
+        "lastMileTravel": 4.2,
+        "serviceability": "SERVICEABLE",
+        "slaString": "27 mins",
+        "lastMileTravelString": "4.2 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 04:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "60% OFF",
+        "subHeader": "UPTO ₹120"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/la-pinoz-pizza-mahagun-mart-gaur-city-2-noida-147910",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "248079",
+      "name": "Subway",
+      "cloudinaryImageId": "1ace5fa65eff3e1223feb696c956b38b",
+      "locality": "Sector 4",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹350 for two",
+      "cuisines": [
+        "Salads",
+        "Snacks",
+        "Desserts",
+        "Beverages"
+      ],
+      "avgRating": 4,
+      "feeDetails": {
+        "restaurantId": "248079",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3900
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3900
+      },
+      "parentId": "2",
+      "avgRatingString": "4.0",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 32,
+        "lastMileTravel": 3,
+        "serviceability": "SERVICEABLE",
+        "slaString": "32 mins",
+        "lastMileTravelString": "3.0 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 04:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/subway-sector-4-gaur-city-1-noida-248079",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "218946",
+      "name": "Wow! Momo",
+      "cloudinaryImageId": "0984acc0ed7b914206dbbcb90297becc",
+      "locality": "Gaur City Mall",
+      "areaName": "Sector 4",
+      "costForTwo": "₹300 for two",
+      "cuisines": [
+        "Tibetan",
+        "Healthy Food",
+        "Asian",
+        "Chinese",
+        "Snacks",
+        "Continental",
+        "Desserts",
+        "Beverages"
+      ],
+      "avgRating": 3.7,
+      "feeDetails": {
+        "restaurantId": "218946",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3900
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3900
+      },
+      "parentId": "1776",
+      "avgRatingString": "3.7",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 36,
+        "lastMileTravel": 3,
+        "serviceability": "SERVICEABLE",
+        "slaString": "36 mins",
+        "lastMileTravelString": "3.0 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "60% OFF",
+        "subHeader": "UPTO ₹120"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/wow-momo-gaur-city-mall-sector-4-noida-218946",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "173374",
+      "name": "Faasos - Wraps & Rolls",
+      "cloudinaryImageId": "af33b81798b11deba338e94b7585d348",
+      "locality": "Gaur Global Village",
+      "areaName": "crossing republic",
+      "costForTwo": "₹200 for two",
+      "cuisines": [
+        "Kebabs",
+        "Fast Food",
+        "Snacks",
+        "North Indian",
+        "American",
+        "Healthy Food",
+        "Desserts",
+        "Beverages"
+      ],
+      "avgRating": 4.1,
+      "feeDetails": {
+        "restaurantId": "173374",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 4500
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 4500
+      },
+      "parentId": "21809",
+      "avgRatingString": "4.1",
+      "totalRatingsString": "5K+",
+      "sla": {
+        "deliveryTime": 31,
+        "lastMileTravel": 4.3,
+        "serviceability": "SERVICEABLE",
+        "slaString": "31 mins",
+        "lastMileTravelString": "4.3 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 03:00:00",
+        "opened": true
+      },
+      "badges": {
+        "textExtendedBadges": [
+          {
+            "iconId": "guiltfree/GF_Logo_android_3x",
+            "shortDescription": "options available",
+            "fontColor": "#7E808C"
+          }
+        ]
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+            "badgeObject": [
+              {
+                "attributes": {
+                  "description": "",
+                  "fontColor": "#7E808C",
+                  "iconId": "guiltfree/GF_Logo_android_3x",
+                  "shortDescription": "options available"
+                }
+              }
+            ]
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "40% OFF",
+        "subHeader": "UPTO ₹80"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/faasos-wraps-and-rolls-gaur-global-village-crossing-republic-noida-173374",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "407887",
+      "name": "The Belgian Waffle Co.",
+      "cloudinaryImageId": "8cb220fa3997a3b928d2ffb6c098acaf",
+      "locality": "Orbit Plaza",
+      "areaName": "crossing republic",
+      "costForTwo": "₹200 for two",
+      "cuisines": [
+        "Waffle",
+        "Desserts",
+        "Ice Cream"
+      ],
+      "avgRating": 4.3,
+      "veg": true,
+      "feeDetails": {
+        "restaurantId": "407887",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 4500
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 4500
+      },
+      "parentId": "2233",
+      "avgRatingString": "4.3",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 36,
+        "lastMileTravel": 4,
+        "serviceability": "SERVICEABLE",
+        "slaString": "36 mins",
+        "lastMileTravelString": "4.0 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 03:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "20% OFF",
+        "subHeader": "UPTO ₹50"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/the-belgian-waffle-co-orbit-plaza-crossing-republic-noida-407887",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "391409",
+      "name": "Burger Singh (Big Punjabi Burgers)",
+      "cloudinaryImageId": "reevvyksx0o9h9o1mmrl",
+      "locality": "Parthala Road",
+      "areaName": "Sector 122",
+      "costForTwo": "₹300 for two",
+      "cuisines": [
+        "Snacks",
+        "Indian",
+        "American",
+        "Beverages",
+        "Desserts"
+      ],
+      "avgRating": 4.3,
+      "feeDetails": {
+        "restaurantId": "391409",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 5700
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 5700
+      },
+      "parentId": "375065",
+      "avgRatingString": "4.3",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 31,
+        "lastMileTravel": 6.6,
+        "serviceability": "SERVICEABLE",
+        "slaString": "31 mins",
+        "lastMileTravelString": "6.6 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 02:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "40% OFF",
+        "subHeader": "UPTO ₹80"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/burger-singh-big-punjabi-burgers-parthala-road-sector-122-noida-391409",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "254002",
+      "name": "McDonald's",
+      "cloudinaryImageId": "6dc3ed2ca21d71acff7c2a51dfe4c720",
+      "locality": "Opulent Mall",
+      "areaName": "Nehru Nagar",
+      "costForTwo": "₹400 for two",
+      "cuisines": [
+        "American"
+      ],
+      "avgRating": 4.2,
+      "feeDetails": {
+        "restaurantId": "254002",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 7700
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 7700
+      },
+      "parentId": "630",
+      "avgRatingString": "4.2",
+      "totalRatingsString": "5K+",
+      "sla": {
+        "deliveryTime": 45,
+        "lastMileTravel": 9,
+        "serviceability": "SERVICEABLE",
+        "slaString": "45 mins",
+        "lastMileTravelString": "9.0 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:25:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "40% OFF",
+        "subHeader": "UPTO ₹80"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/mcdonalds-opulent-mall-nehru-nagar-noida-254002",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "213470",
+      "name": "Taco Bell",
+      "cloudinaryImageId": "d3b3db238b6448c3f297c851e9d0b96b",
+      "locality": "Sector 4",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹300 for two",
+      "cuisines": [
+        "Mexican"
+      ],
+      "avgRating": 4,
+      "feeDetails": {
+        "restaurantId": "213470",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3300
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3300
+      },
+      "parentId": "1557",
+      "avgRatingString": "4.0",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 36,
+        "lastMileTravel": 2.7,
+        "serviceability": "SERVICEABLE",
+        "slaString": "36 mins",
+        "lastMileTravelString": "2.7 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 03:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "40% OFF",
+        "subHeader": "UPTO ₹80"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/taco-bell-sector-4-gaur-city-1-noida-213470",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "245432",
+      "name": "Chai Point",
+      "cloudinaryImageId": "pt1qslmxiudsjdjufvaz",
+      "locality": "Gaur City Mall",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹150 for two",
+      "cuisines": [
+        "Bakery",
+        "Beverages",
+        "Maharashtrian",
+        "Snacks",
+        "Street Food",
+        "South Indian",
+        "Punjabi",
+        "Chaat",
+        "Indian",
+        "American",
+        "North Indian",
+        "Fast Food",
+        "Desserts",
+        "Cafe",
+        "Healthy Food",
+        "Home Food"
+      ],
+      "avgRating": 4.1,
+      "feeDetails": {
+        "restaurantId": "245432",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3300
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3300
+      },
+      "parentId": "1607",
+      "avgRatingString": "4.1",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 30,
+        "lastMileTravel": 2.7,
+        "serviceability": "SERVICEABLE",
+        "slaString": "30 mins",
+        "lastMileTravelString": "2.7 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 22:00:00",
+        "opened": true
+      },
+      "badges": {
+        "textExtendedBadges": [
+          {
+            "iconId": "guiltfree/GF_Logo_android_3x",
+            "shortDescription": "options available",
+            "fontColor": "#7E808C"
+          }
+        ]
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+            "badgeObject": [
+              {
+                "attributes": {
+                  "description": "",
+                  "fontColor": "#7E808C",
+                  "iconId": "guiltfree/GF_Logo_android_3x",
+                  "shortDescription": "options available"
+                }
+              }
+            ]
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "60% OFF",
+        "subHeader": "UPTO ₹120"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/chai-point-gaur-city-mall-gaur-city-1-noida-245432",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "173378",
+      "name": "Oven Story Pizza",
+      "cloudinaryImageId": "f986df6f1a1fcf2ff24d2eaf44d570a7",
+      "locality": "Gaur Global Village",
+      "areaName": "crossing republic",
+      "costForTwo": "₹400 for two",
+      "cuisines": [
+        "Pizzas",
+        "Pastas",
+        "Italian",
+        "Desserts",
+        "Beverages"
+      ],
+      "avgRating": 3.6,
+      "feeDetails": {
+        "restaurantId": "173378",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 5000
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 5000
+      },
+      "parentId": "3534",
+      "avgRatingString": "3.6",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 36,
+        "lastMileTravel": 4.3,
+        "serviceability": "SERVICEABLE",
+        "slaString": "30-42 mins",
+        "lastMileTravelString": "4.3 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 03:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "60% OFF",
+        "subHeader": "UPTO ₹120"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/oven-story-pizza-gaur-global-village-crossing-republic-noida-173378",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "579096",
+      "name": "EatFit",
+      "cloudinaryImageId": "6126c9b45de2cb222405c1af8a321e74",
+      "locality": "C Block",
+      "areaName": "Sector 63",
+      "costForTwo": "₹270 for two",
+      "cuisines": [
+        "Chinese",
+        "Healthy Food",
+        "Tandoor",
+        "Pizzas",
+        "North Indian",
+        "Thalis",
+        "Biryani"
+      ],
+      "avgRating": 4.1,
+      "feeDetails": {
+        "restaurantId": "579096",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 6400
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 6400
+      },
+      "parentId": "76139",
+      "avgRatingString": "4.1",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 34,
+        "lastMileTravel": 7.7,
+        "serviceability": "SERVICEABLE",
+        "slaString": "34 mins",
+        "lastMileTravelString": "7.7 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:59:00",
+        "opened": true
+      },
+      "badges": {
+        "textExtendedBadges": [
+          {
+            "iconId": "guiltfree/GF_Logo_android_3x",
+            "shortDescription": "brand",
+            "fontColor": "#7E808C"
+          }
+        ]
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+            "badgeObject": [
+              {
+                "attributes": {
+                  "description": "",
+                  "fontColor": "#7E808C",
+                  "iconId": "guiltfree/GF_Logo_android_3x",
+                  "shortDescription": "brand"
+                }
+              }
+            ]
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "60% OFF",
+        "subHeader": "UPTO ₹120"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/eatfit-c-block-sector-63-noida-579096",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "427017",
+      "name": "Baskin Robbins - Ice Cream Desserts",
+      "cloudinaryImageId": "85ccae4e3576f9330af102c46ca85395",
+      "locality": "Sector 4",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹250 for two",
+      "cuisines": [
+        "Desserts",
+        "Ice Cream"
+      ],
+      "avgRating": 4.3,
+      "veg": true,
+      "feeDetails": {
+        "restaurantId": "427017",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3300
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3300
+      },
+      "parentId": "5588",
+      "avgRatingString": "4.3",
+      "totalRatingsString": "100+",
+      "sla": {
+        "deliveryTime": 25,
+        "lastMileTravel": 2.7,
+        "serviceability": "SERVICEABLE",
+        "slaString": "25 mins",
+        "lastMileTravelString": "2.7 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:59:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "20% OFF",
+        "subHeader": "UPTO ₹50"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/baskin-robbins-ice-cream-desserts-sector-4-gaur-city-1-noida-427017",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "230304",
+      "name": "Keventers - Milkshakes & Desserts",
+      "cloudinaryImageId": "f0e94c984813053f46c214f18ce73d59",
+      "locality": "Gaur City Mall",
+      "areaName": "Sector 4",
+      "costForTwo": "₹200 for two",
+      "cuisines": [
+        "Beverages",
+        "Ice Cream",
+        "Desserts",
+        "Healthy Food"
+      ],
+      "avgRating": 3.7,
+      "veg": true,
+      "feeDetails": {
+        "restaurantId": "230304",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3300
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3300
+      },
+      "parentId": "268997",
+      "avgRatingString": "3.7",
+      "totalRatingsString": "500+",
+      "sla": {
+        "deliveryTime": 26,
+        "lastMileTravel": 2.7,
+        "serviceability": "SERVICEABLE",
+        "slaString": "26 mins",
+        "lastMileTravelString": "2.7 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:59:00",
+        "opened": true
+      },
+      "badges": {
+        "textExtendedBadges": [
+          {
+            "iconId": "guiltfree/GF_Logo_android_3x",
+            "shortDescription": "options available",
+            "fontColor": "#7E808C"
+          }
+        ]
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+            "badgeObject": [
+              {
+                "attributes": {
+                  "description": "",
+                  "fontColor": "#7E808C",
+                  "iconId": "guiltfree/GF_Logo_android_3x",
+                  "shortDescription": "options available"
+                }
+              }
+            ]
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "40% OFF",
+        "subHeader": "UPTO ₹80"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/keventers-milkshakes-and-desserts-gaur-city-mall-sector-4-noida-230304",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "220495",
+      "name": "Mad Over Donuts",
+      "cloudinaryImageId": "b54fb45b7a9608cc851d1e3702436497",
+      "locality": "Gaur City Mall",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹200 for two",
+      "cuisines": [
+        "Desserts",
+        "Sweets",
+        "Bakery"
+      ],
+      "avgRating": 4.3,
+      "feeDetails": {
+        "restaurantId": "220495",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3300
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3300
+      },
+      "parentId": "611",
+      "avgRatingString": "4.3",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 24,
+        "lastMileTravel": 2.7,
+        "serviceability": "SERVICEABLE",
+        "slaString": "24 mins",
+        "lastMileTravelString": "2.7 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 23:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "40% OFF",
+        "subHeader": "UPTO ₹80"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/mad-over-donuts-gaur-city-mall-gaur-city-1-noida-220495",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "352737",
+      "name": "UBQ BY Barbeque Nation",
+      "cloudinaryImageId": "gnoatstgumljssdj1qxu",
+      "locality": "Sector 4",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹300 for two",
+      "cuisines": [
+        "North Indian",
+        "Barbecue",
+        "Biryani",
+        "Kebabs",
+        "Mughlai",
+        "Desserts"
+      ],
+      "avgRating": 3.5,
+      "feeDetails": {
+        "restaurantId": "352737",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 4400
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 4400
+      },
+      "parentId": "10804",
+      "avgRatingString": "3.5",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 35,
+        "lastMileTravel": 3,
+        "serviceability": "SERVICEABLE",
+        "slaString": "35 mins",
+        "lastMileTravelString": "3.0 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-07 16:30:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "40% OFF",
+        "subHeader": "UPTO ₹80"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/ubq-by-barbeque-nation-sector-4-gaur-city-1-noida-352737",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "252613",
+      "name": "NIC Ice Creams",
+      "cloudinaryImageId": "c8b98c4064dc2fd6bdf350212e5f0bd5",
+      "locality": "Biharipur Village",
+      "areaName": "crossing republic",
+      "costForTwo": "₹120 for two",
+      "cuisines": [
+        "Ice Cream",
+        "Desserts"
+      ],
+      "avgRating": 4.6,
+      "veg": true,
+      "feeDetails": {
+        "restaurantId": "252613",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3900
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3900
+      },
+      "parentId": "6249",
+      "avgRatingString": "4.6",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 24,
+        "lastMileTravel": 3,
+        "serviceability": "SERVICEABLE",
+        "slaString": "24 mins",
+        "lastMileTravelString": "3.0 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 00:00:00",
+        "opened": true
+      },
+      "badges": {
+
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "EVERY ITEM",
+        "subHeader": "@ ₹159"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/nic-ice-creams-biharipur-village-crossing-republic-noida-252613",
+      "type": "WEBLINK"
+    }
+  },
+  {
+    "info": {
+      "id": "236299",
+      "name": "Biryani Blues",
+      "cloudinaryImageId": "gjt1cwtxijmrsphebwks",
+      "locality": "Sector 4",
+      "areaName": "Gaur City 1",
+      "costForTwo": "₹400 for two",
+      "cuisines": [
+        "Biryani",
+        "Hyderabadi",
+        "Lucknowi",
+        "Kebabs",
+        "Desserts",
+        "Beverages"
+      ],
+      "avgRating": 4.1,
+      "feeDetails": {
+        "restaurantId": "236299",
+        "fees": [
+          {
+            "name": "BASE_DISTANCE",
+            "fee": 3300
+          },
+          {
+            "name": "BASE_TIME"
+          },
+          {
+            "name": "ANCILLARY_SURGE_FEE"
+          }
+        ],
+        "totalFee": 3300
+      },
+      "parentId": "13813",
+      "avgRatingString": "4.1",
+      "totalRatingsString": "1K+",
+      "sla": {
+        "deliveryTime": 28,
+        "lastMileTravel": 2.7,
+        "serviceability": "SERVICEABLE",
+        "slaString": "28 mins",
+        "lastMileTravelString": "2.7 km",
+        "iconType": "ICON_TYPE_EMPTY"
+      },
+      "availability": {
+        "nextCloseTime": "2023-09-08 00:00:00",
+        "opened": true
+      },
+      "badges": {
+        "textExtendedBadges": [
+          {
+            "iconId": "guiltfree/GF_Logo_android_3x",
+            "shortDescription": "options available",
+            "fontColor": "#7E808C"
+          }
+        ]
+      },
+      "isOpen": true,
+      "type": "F",
+      "badgesV2": {
+        "entityBadges": {
+          "imageBased": {
+
+          },
+          "textBased": {
+
+          },
+          "textExtendedBadges": {
+            "badgeObject": [
+              {
+                "attributes": {
+                  "description": "",
+                  "fontColor": "#7E808C",
+                  "iconId": "guiltfree/GF_Logo_android_3x",
+                  "shortDescription": "options available"
+                }
+              }
+            ]
+          }
+        }
+      },
+      "aggregatedDiscountInfoV3": {
+        "header": "50% OFF",
+        "subHeader": "UPTO ₹100"
+      },
+      "loyaltyDiscoverPresentationInfo": {
+        "logoCtx": {
+          "text": "BENEFITS",
+          "logo": "v1634558776/swiggy_one/OneLogo_3x.png"
+        },
+        "freedelMessage": "FREE DELIVERY"
+      },
+      "differentiatedUi": {
+        "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        "differentiatedUiMediaDetails": {
+          "mediaType": "ADS_MEDIA_ENUM_IMAGE",
+          "lottie": {
+
+          },
+          "video": {
+
+          }
+        }
+      },
+      "reviewsSummary": {
+
+      },
+      "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      "restaurantOfferPresentationInfo": {
+
+      }
+    },
+    "analytics": {
+
+    },
+    "cta": {
+      "link": "https://www.swiggy.com/restaurants/biryani-blues-sector-4-gaur-city-1-noida-236299",
+      "type": "WEBLINK"
+    }
+  }
+]
+
+const BodyComponent = () => {
+  return (
+    <div className="body">
+      <div className="search-bar">
+        <input type="text" placeholder="Search for restaurants" />
+        <button className="search-button">Search</button>
+      </div>
+      <div className="restaurant-container">
+        {
+          resObjArray.map((resObj) => (
+            < RestaurantCardComponent key={resObj.info.id} resData={resObj} />
+          ))
+        }
+      </div>
+    </div>
+  );
+};
+
+export default BodyComponent;
