+++ src/components/FooterComponent.js
@@ -1,0 +1,3 @@
+const FooterComponent = () => { };
+
+export default FooterComponent;
